+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## Unreleased
 ### Changed
 - Cleaned up app, root panels and tab bar, prepare for UIUC 4 features [#1629](https://github.com/rokwire/illinois-app/issues/1629).
+### Added
+- Created Favorites root panel [#1629](https://github.com/rokwire/illinois-app/issues/1629).
 
 ## [4.0.9] - 2022-05-13
 ### Changed